Switch to using VARANT as the drawing number.  drawing number now only refers to the index of a set of drawings on the view

+++ openPER/ViewModels/PartViewModel.cs
@@ -5,7 +5,7 @@ namespace openPER.ViewModels
 {
     public class PartViewModel
     {
-        public decimal PartNumber { get; set; }
+        public int PartNumber { get; set; }
         public string Description { get; set; }
         public string FurtherDescription { get; set; }
         public string Quantity { get; set; }
